requirements.txt
arrow
datasets
einops
python-dotenv
pandas
scikit-learn
scipy
tqdm
umap-learn
seaborn
h5py
opencv-python
pytest-env
dagshub
httpx
#faiss
gdown
transformers
accelerate
ipdb
ipykernel
ipython
jupyter-client
jupyter-core
jupyter-events
jupyter-lsp
jupyter-server
jupyter-server-terminals
jupyterlab
jupyterlab-pygments
jupyterlab-server
lightning
lightning-utilities
mlflow
hydra-core
omegaconf
scikit-dimension
kneed
black
isort
git+https://github.com/HamidrezaKmK/nflows.git
matplotlib
numpy
pillow
torch
torchvision
